add unitTest
